feat: Do not waste bandwidth - do not recreate page on hover actions (#2)
* feat: Do not redraw on mouse over. Try to get the hover/css stuff working.

* feat: Progress.

* fix: Fix everything.

---------

+++ src/DayAheadPrice/Entities/PriceList.cs
@@ -9,7 +9,7 @@ public class PriceList
     /// <summary>
     /// List of prices per hour.
     /// </summary>
-    public SortedList<DateTime, double> Prices { get; set; } = new();
+    public SortedList<DateTime, decimal> Prices { get; set; } = new();
 
     /// <summary>
     /// Gets electricity prices adjusted by VAT and operator margin.
@@ -17,13 +17,13 @@ public class PriceList
     /// <param name="margin">Operator margin (c/kWh).</param>
     /// <param name="vat">VAT (simple multiplier, 25.5% => 0.255).</param>
     /// <returns>Sorted list with time and electricity price values (c/kWh).</returns>
-    public SortedList<DateTime, double> GetPrices(decimal margin, decimal vat)
+    public SortedList<DateTime, decimal> GetPrices(decimal margin, decimal vat)
     {
-        var adjustedPrices = new SortedList<DateTime, double>();
+        var adjustedPrices = new SortedList<DateTime, decimal>();
 
         foreach (var price in Prices)
         {
-            adjustedPrices.Add(price.Key, price.Value * (double)(1m + vat) + (double)margin);
+            adjustedPrices.Add(price.Key, price.Value * (1m + vat) + margin);
         }
 
         return adjustedPrices;
